social_media_service/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
drf-spectacular==0.28.0
django-jazzmin==3.0.1
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
Pillow==11.1.0
